1.建立AirMainActivity使用navigationtabbar下方滑動搭配Fragment(AQI、PM2.5、AIR) 2.AQIFragment設計完成(尚未連棟資料)，使用DashboardView計量表顯示數值

+++ app/src/main/java/com/openweather/openweather/LoadingSplash/SplashActivity.java
@@ -488,9 +488,9 @@ public class SplashActivity extends AppCompatActivity  {
                             cl5.moveToFirst();
                             if(cl5.getCount()==0) {
                                 mAccess2.add();
-                                mAccess2.add("1",Integer.parseInt(UVI),PublishAgency,PublishTime,SiteName,mLatitude,mLongitude);
+                                mAccess2.add("1",(int)Double.parseDouble(UVI),PublishAgency,PublishTime,SiteName,mLatitude,mLongitude);
                             }else{
-                                mAccess2.update("1",Integer.parseInt(UVI),PublishAgency,PublishTime,SiteName,mLatitude,mLongitude,null);
+                                mAccess2.update("1",(int)Double.parseDouble(UVI),PublishAgency,PublishTime,SiteName,mLatitude,mLongitude,null);
                             }
                         } catch (JSONException e) {
                             e.printStackTrace();
